settings: make RegisterEnumSetting generic
We change `RegisterEnumSetting` to use any integer type instead of
always `int64`. This makes the code safer and cleaner (we no longer
need to/from `int64` casts).

Epic: none
Release note: None

+++ pkg/ccl/kvccl/kvtenantccl/upgradeinterlockccl/local_test_util_test.go
@@ -81,7 +81,7 @@ func runTest(t *testing.T, variant sharedtestutil.TestVariant, test sharedtestut
 		stats.AutomaticStatisticsClusterMode.Override(ctx, &s.SV, false)
 		stats.UseStatisticsOnSystemTables.Override(ctx, &s.SV, false)
 		stats.AutomaticStatisticsOnSystemTables.Override(ctx, &s.SV, false)
-		sql.DistSQLClusterExecMode.Override(ctx, &s.SV, int64(sessiondatapb.DistSQLOff))
+		sql.DistSQLClusterExecMode.Override(ctx, &s.SV, sessiondatapb.DistSQLOff)
 	}
 
 	reduceLeaseDurationAndReclaimLoopInterval := func(s *cluster.Settings) {
